+++ src/Bpp/Phyl/Model/Protein/Coala.cpp
@@ -161,7 +161,7 @@ void Coala::computeEquilibriumFrequencies()
     if (norm == true)
     {
       double s = VectorTools::sum(freq_);
-      for (int i = 0; i < 20; i++)
+      for (size_t i = 0; i < 20; i++)
       {
         freq_[i] = freq_[i] / s;
       }
